🪄 更新检查、文件检查
