--- ToDo.md
@@ -1,37 +1,0 @@
-
-## 功能计划清单
-
-### 核心测试
-
-- [x] 配置文件完整性检查
-- [ ] 数据库结构完整性检查
-
-### 后端 Bot 侧
-
-- [x] 启动时间记录
-- [x] 上线提醒
-- [x] 好友/邀群申请
-
-### 后端 API 侧
-
-- [x] 写配置文件
-- [x] 面板总览数据接口，分钟轮询
-- [x] 入群自动手动切换
-- [x] 申请列表五秒轮询
-- [x] 直播间列表、联系人列表切换更新
-- [x] 申请项目查询、操作
-- [x] 直播间绑定项目查询、关注、取关
-- [x] 群发
-- [x] 直播检测工作模式切换
-- [x] 直播检测账号登录
-- [x] 联系人项目查询、操作
-- [x] QQ、Q群、B站等资源跨域转发
-
-### 前端
-
-- [x] 基础面板
-- [x] 基础逻辑
-- [x] 配置文件对话框
-- [x] B站账号登录对话框
-- [x] 直播间绑定对话框
-- [x] 联系人项目对话框
