+++ common/preload.js
@@ -3,12 +3,18 @@
  */
 
 const logger = require('npmlog');
-
+const fs = require('fs');
+const path = require('path');
 const axios = require('axios');
 
+const { version } = require('../package.json');
 const { configDB } = require('../db');
 
 module.exports = async function () {
+
+    // 获取版本信息
+    const versionInfo = await axios(`https://yinlan-bot.oss-cn-beijing.aliyuncs.com/livebot/version/${version}/version.json`).then(resp => resp.data);
+    fs.writeFileSync(path.resolve('temp', 'version.json'), JSON.stringify(versionInfo));
 
     // 读取环境变量
     if (process.env.NODE_ENV == 'development') process.env.dev = 1;
@@ -16,7 +22,6 @@ module.exports = async function () {
         logger.info("当前为开发环境");
         logger.warn("开发环境下身份验证将被禁用");
     }
-  
 
     // 清空状态数据库
     configDB.prepare(`UPDATE statusConfig SET value = 0;`).run();
