+++ express/routes/index.js
@@ -2,21 +2,22 @@
 // 引入库
 const express = require('express');
 
-const { version } = require('../../package.json');
+const { version, publish } = require('../../temp/version.json');
 
 const config = require('../../config.json');
 
 // 初始化路由
 const router = express.Router();
 
-router.get('/', (req, res) => {
+router.get('/', async (req, res) => {
 
-    let status = config.link ? 'active' : 'uninitialized';
-
+    const status = config.link ? 'active' : 'uninitialized';
+    
     res.send({
         status,
         yinlan: 'livebot',
         version,
+        publish,
         uwu: 'Made with ♥ by colour93',
         doc: 'https://livebot.yinlan.furbot.icu',
         repo: 'https://github.com/colour93/yinlan-livebot'
